References to rowid should be 64bit
This is due to SQLite treating the ROWID value as an 64bit integer.

+++ collector/src/model/record.rs
@@ -3,8 +3,8 @@ use super::super::schema::record;
 #[derive(Queryable)]
 pub struct Record {
     pub rowid: i64,
-    pub domain: i32,
-    pub parent: i32,
+    pub domain: i64,
+    pub parent: i64,
     pub response_code: i32,
     pub record_type: String,
     pub ttl: i32,
@@ -16,8 +16,8 @@ pub struct Record {
 #[derive(Insertable)]
 #[table_name = "record"]
 pub struct NewRecord<'a> {
-    pub domain: &'a i32,
-    pub parent: &'a i32,
+    pub domain: &'a i64,
+    pub parent: &'a i64,
     pub response_code: &'a i32,
     pub record_type: &'a str,
     pub ttl: &'a i32,
